+++ source/matrixes.cpp
@@ -1,12 +1,12 @@
 #include "matrixes.h"
 
 TridiagonalMatrix::TridiagonalMatrix(
-    const std::vector<std::array<float, 3>> &data)
+    const std::vector<std::array<double, 3>> &data)
     : m_data{data} {}
 
 int TridiagonalMatrix::get_size() const { return m_data.size(); }
 
-float TridiagonalMatrix::operator()(int i, int j) const {
+double TridiagonalMatrix::operator()(int i, int j) const {
     if (j == i - 1) {
         return m_data[i][0];
     } else if (j == i) {
@@ -18,7 +18,7 @@ float TridiagonalMatrix::operator()(int i, int j) const {
     }
 }
 
-CSRMatrix::CSRMatrix(const std::map<std::pair<int, int>, float> &data) {
+CSRMatrix::CSRMatrix(const std::map<std::pair<int, int>, double> &data) {
     m_values.reserve(data.size());
     m_column_indexes.reserve(data.size());
     m_row_indexation.reserve(data.size());
@@ -37,7 +37,7 @@ CSRMatrix::CSRMatrix(const std::map<std::pair<int, int>, float> &data) {
     }
 }
 
-const std::vector<float> &CSRMatrix::get_values() const { return m_values; }
+const std::vector<double> &CSRMatrix::get_values() const { return m_values; }
 const std::vector<int> &CSRMatrix::get_column_indexes() const {
     return m_column_indexes;
 }
@@ -45,7 +45,7 @@ const std::vector<int> &CSRMatrix::get_row_indexation() const {
     return m_row_indexation;
 }
 
-float CSRMatrix::operator()(int i, int j) const {
+double CSRMatrix::operator()(int i, int j) const {
     if (i + 1 >= m_column_indexes.size()) {
         return 0;
     }
@@ -57,12 +57,12 @@ float CSRMatrix::operator()(int i, int j) const {
     return 0;
 }
 
-std::vector<float>
-CSRMatrix::operator*(const std::vector<float> &column) const {
-    std::vector<float> res;
+std::vector<double>
+CSRMatrix::operator*(const std::vector<double> &column) const {
+    std::vector<double> res;
     res.reserve(m_row_indexation.size() - 1);
     for (int k = 0; k < m_row_indexation.size() - 1; ++k) {
-        float temp = 0;
+        double temp = 0;
         for (int i = m_row_indexation[k]; i < m_row_indexation[k + 1]; ++i) {
             temp += m_values[i] * column[m_column_indexes[i]];
         }
@@ -71,7 +71,7 @@ CSRMatrix::operator*(const std::vector<float> &column) const {
     return res;
 }
 
-DenseMatrix::DenseMatrix(const std::vector<float> &data, int columns_num)
+DenseMatrix::DenseMatrix(const std::vector<double> &data, int columns_num)
     : m_data{data}, m_columns{columns_num}, m_rows{(data.size() + columns_num -
                                                     1) /
                                                    columns_num} {
@@ -82,7 +82,7 @@ int DenseMatrix::get_number_of_columns() const { return m_columns; }
 
 int DenseMatrix::get_number_of_rows() const { return m_rows; }
 
-float DenseMatrix::operator()(int i, int j) const {
+double DenseMatrix::operator()(int i, int j) const {
     if (i * m_columns + j >= m_data.size()) {
         return 0;
     }
@@ -100,12 +100,12 @@ bool DenseMatrix::operator==(const DenseMatrix &other) const {
     return true;
 }
 
-void DenseMatrix::set_item(int i, int j, float value) {
+void DenseMatrix::set_item(int i, int j, double value) {
     m_data[i * m_columns + j] = value;
 }
 
-std::vector<float> DenseMatrix::get_column(int i) const {
-    std::vector<float> res;
+std::vector<double> DenseMatrix::get_column(int i) const {
+    std::vector<double> res;
     res.reserve(m_rows);
     for (int j = 0; j < m_rows; ++j) {
         res.push_back((*this)(j, i));
@@ -113,8 +113,8 @@ std::vector<float> DenseMatrix::get_column(int i) const {
     return res;
 }
 
-std::vector<float> DenseMatrix::get_row(int i) const {
-    std::vector<float> res;
+std::vector<double> DenseMatrix::get_row(int i) const {
+    std::vector<double> res;
     res.reserve(m_columns);
     for (int j = 0; j < m_columns; ++j) {
         res.push_back((*this)(i, j));
@@ -122,12 +122,12 @@ std::vector<float> DenseMatrix::get_row(int i) const {
     return res;
 }
 
-std::vector<float>
-DenseMatrix::operator*(const std::vector<float> &column) const {
+std::vector<double>
+DenseMatrix::operator*(const std::vector<double> &column) const {
     assert(m_columns == column.size() &&
            "the matrix's number of columns must be equal to column size");
-    float temp;
-    std::vector<float> res;
+    double temp;
+    std::vector<double> res;
     for (int i = 0; i < m_rows; ++i) {
         temp = 0;
         for (int j = 0; j < m_columns; ++j) {
