+++ source/solvers.h
@@ -3,7 +3,7 @@
 #ifndef SOLVERS_H
 #define SOLVERS_H
 
-std::vector<float> solve(const TridiagonalMatrix &matrix,
-                         const std::vector<float> &column);
+std::vector<double> solve(const TridiagonalMatrix &matrix,
+                          const std::vector<double> &column);
 
 #endif
